get rid of SkeletonTracingController and don't pass SceneController where it's not needed (#1858)
